♻️  Refactor: category 컴포넌트 분리
